Change HibernateMode from "Auto" to "Nvram"
Hey @EETagent,

there is a kind of error in your latest update. `HibernateMode` should be set to `NVRAM` instead of `Auto` as we are blocking out the upper memory bank of the RTC-Memory (80-AB), which is used by OSX for information about hibernation, anyway. The same bank is used and checksummed by Lenovo for diagnostic and internal informations, which led to the CMOS-corruption-errors in first place. So setting this to `Auto` introduces a potential bug and has no value.